pyproject.toml
[tool.poetry]
name = "demo-django"
version = "0.1.0"
description = ""
authors = ["syuji-higa <syuji-higa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1"
psycopg2-binary = "^2.8.5"
djangorestframework = "^3.11.1"
django-cors-headers = "^3.4.0"
drf-yasg = "^1.17.1"
django-3-jet = "^1.0.8"
djangorestframework-camel-case = "^1.2.0"
django-admin-sortable = "^2.2.3"
factory_boy = "^3.0.1"
gunicorn = "^20.0.4"
django-environ = "^0.4.5"
google-cloud-secret-manager = "1.0.0"
google-auth = "^1.21.2"
django-storages = {extras = ["google"], version = "^1.10.1"}

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
black = "^19.10b0"
mypy = "^0.782"
isort = "^5.4.2"
pre-commit = "^2.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
